[CDE-280] - Added IUrlProvider implementations that can be override on cdf and cde if needed. Fixed pluginRepository file handling by packager. Applied check style on changed files
