+++ cpf-pentaho/src/pt/webdetails/cpf/PentahoPluginEnvironment.java
@@ -4,9 +4,7 @@ import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.pentaho.platform.engine.core.system.PentahoRequestContextHolder;
 import org.pentaho.platform.engine.core.system.PentahoSessionHolder;
-import org.pentaho.platform.engine.core.system.PentahoSystem;
 
 import pt.webdetails.cpf.context.api.IUrlProvider;
 import pt.webdetails.cpf.plugin.CorePlugin;
@@ -22,6 +20,7 @@ import pt.webdetails.cpf.repository.util.RepositoryHelper;
 public class PentahoPluginEnvironment extends PentahoBasePluginEnvironment implements IContentAccessFactory {
 
   private static final PentahoPluginEnvironment instance = new PentahoPluginEnvironment();
+  private IUrlProvider pentahoUrlProvider;
 
   static {
     PluginEnvironment.init( instance );
@@ -48,58 +47,13 @@ public class PentahoPluginEnvironment extends PentahoBasePluginEnvironment imple
     return new PluginLegacySolutionResourceAccess( basePath );
   }
 
-
   @Override
   public IUrlProvider getUrlProvider() {
-    return new IUrlProvider() {
-
-      private static final String CONTENT = "content";
-
-      @Override
-      public String getPluginStaticBaseUrl() {
-        return getPluginStaticBaseUrl( getPluginId() );
-      }
-
-      @Override
-      public String getPluginStaticBaseUrl( String pluginId ) {
-        return Util.joinPath( getApplicationBaseUrl(), CONTENT, pluginId ) + "/";
-      }
-
-      @Override
-      public String getPluginBaseUrl() {
-        return getPluginBaseUrl( getPluginId() );
-      }
-
-      @Override
-      public String getPluginBaseUrl( String pluginId ) {
-        return Util.joinPath( getApplicationBaseUrl(), CONTENT, pluginId ) + "/";
-      }
-
-      @SuppressWarnings( "deprecation" )
-      protected String getApplicationBaseUrl() {
-        return PentahoSystem.getApplicationContext().getBaseUrl();
-      }
-
-      @Override
-      public String getRepositoryUrl( String fullPath ) {
-        return Util.joinPath( getApplicationBaseUrl(), CONTENT, getPluginId(), "res", fullPath );
-      }
-
-      @Override
-      public String getWebappContextPath() {
-        return PentahoRequestContextHolder.getRequestContext().getContextPath();
-      }
-
-      @Override
-      public String getWebappContextRoot() {
-        String url = PentahoSystem.getApplicationContext().getFullyQualifiedServerURL(),
-          webappName = getWebappContextPath();
-
-        return url.substring( 0, url.length() - webappName.length() + 1 );
-      }
-    };
+    if( pentahoUrlProvider == null ) {
+      pentahoUrlProvider = new PentahoUrlProvider( getPluginId() );
+    }
+    return pentahoUrlProvider;
   }
-
 
   //FIXME
   public IPluginCall getPluginCall( String pluginId, String servicePath, String method ) {
